More multiplayer stuff removed
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8170 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ sound/snd_decoder.cpp
@@ -499,7 +499,7 @@ int idSampleDecoderLocal::DecodePCM( idSoundSample *sample, int sampleOffset44k,
 		return 0;
 	}
 
-	if ( size - pos < sampleCount * sizeof( short ) ) {
+	if ( size - pos < sampleCount * (int)sizeof( short ) ) {
 		readSamples = ( size - pos ) / sizeof( short );
 	} else {
 		readSamples = sampleCount;
